Moved ``indent`` to ``utility.py`` (#180)
* moved indent to utility

* others indent imports
