+++ fast64_internal/oot/oot_utility.py
@@ -1,12 +1,6 @@
 import bpy, math, os, re
 from bpy.utils import register_class, unregister_class
 from .oot_constants import ootSceneIDToName
-
-
-def isPathObject(obj: bpy.types.Object) -> bool:
-    return obj.data is not None and isinstance(obj.data, bpy.types.Curve) and obj.ootSplineProperty.splineType == "Path"
-
-
 from ..utility import (
     PluginError,
     prop_split,
@@ -20,8 +14,8 @@ from ..utility import (
     hexOrDecInt,
 )
 
-# default indentation to use when writing to decomp files
-indent = " " * 4
+def isPathObject(obj: bpy.types.Object) -> bool:
+    return obj.data is not None and isinstance(obj.data, bpy.types.Curve) and obj.ootSplineProperty.splineType == "Path"
 
 ootSceneDungeons = [
     "bdan",
