+++ fast64_internal/oot/scene/exporter/to_c/room_header.py
@@ -1,5 +1,4 @@
-from .....utility import CData
-from ....oot_utility import indent
+from .....utility import CData, indent
 from ....oot_level_classes import OOTRoom, OOTScene
 from .actor import ootActorListToC
 from .room_commands import ootRoomCommandsToC
